requirements.txt
# --- Streamlit & Utils ---
streamlit>=1.31.0
python-dotenv>=1.0.0
backoff>=2.2.0

# --- LlamaIndex (Core, Readers, Embeddings, Groq) ---
llama-index-core>=0.12.0
llama-index-readers-file>=0.4.0
llama-index-embeddings-huggingface>=0.5.0
llama-index-llms-groq>=0.3.0

# --- Transformers ---
sentence-transformers>=3.0.0

# --- LLM API ---
groq>=0.9.0

# --- Excel/CSV Support ---
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
xlwt>=1.3.0

# --- Word/PDF Support ---
python-docx>=0.8.11
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# --- Table Extraction from PDF ---
camelot-py[cv]
ghostscript

# --- OCR/Images for PDF ---
pytesseract
pdf2image
Pillow
poppler-utils

# --- END ---
